inc_api/requirements.txt
Django==4.0.4
djangorestframework==3.14.0
SQLAlchemy==1.4.26
django-role-permissions==3.1.1
whitenoise==5.3.0
psycopg2==2.9.1
gunicorn==20.1.0
Markdown==3.3.4
django-filter==22.1
django-phonenumber-field==7.0.0
django-cors-headers==3.13.0
Pillow==8.4.0
XlsxWriter==3.0.3